This change refs #56153
The $(MAKEFLAGS) special variable was misused in the top-level makefile to pass down custom options.  This breaks the capability of using the FV3/makefile as a submake from a larger build system.  The new NEMS build system is entirely makefiles, and calls FV3/makefile via $(MAKE), so this fix is needed before the NEMS commit.